fix: address code review feedback - move sorting to controller, fix wildcard validation, cast JS counters to int
Agent-Logs-Url: https://github.com/sherwin-cell/jobease/sessions/cc233135-6f8d-4d00-b481-974cd472b137

+++ resources/views/jobseeker/profile/create.blade.php
@@ -307,9 +307,9 @@ document.addEventListener('DOMContentLoaded', () => {
     document.querySelectorAll('.current-job-checkbox').forEach(bindCurrentJobCheckbox);
 
     // ── Counters (avoid index collision after removes) ──
-    let expCount  = {{ count($experiences ?? []) }};
-    let eduCount  = {{ count($educations ?? []) }};
-    let certCount = {{ count($certs ?? []) }};
+    let expCount  = {{ (int) count($experiences ?? []) }};
+    let eduCount  = {{ (int) count($educations ?? []) }};
+    let certCount = {{ (int) count($certs ?? []) }};
 
     // ── Add Skill ──
     document.getElementById('add-skill').addEventListener('click', () => {
